+++ packages/google_sign_in/google_sign_in_ios/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 6.3.6
 
 * Converts the Pigeon host API from Objective-C to Swift.
+* Converts `SignInResult` to a sealed Pigeon type.
 
 ## 6.3.5